+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add options to send Webmin and Usermin errors to the systemd journal [forum.virtualmin.com/t/136562](https://forum.virtualmin.com/t/miniserv-webserver-log-growing-too-big-should-be-rotated/136562)
 * Add webserver logging controls to Usermin Configuration module
 * Add option to rotate Webmin and Usermin webserver logs using `logrotate` instead of periodically clearing them [#2821](https://github.com/webmin/webmin/pull/2821)
+* Add DNF 4 and 5 package hold management to the Software Package Updates module
+* Fix DNF update confirmations by previewing packages and dependencies that will be installed or updated
 * Fix PostgreSQL initialization on EL systems to use SCRAM-SHA-256 authentication by default
 
 #### 2.660 (August 20, 2026)